+++ src/main/java/com/example/appprueba/domain/model/Price.java
@@ -18,7 +18,7 @@ public class Price {
     private Long brandId;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
-    private BigDecimal priceList;
+    private Long priceList;
     private Long productId;
     private Integer priority;
     private BigDecimal price;
